requirements.txt
streamlit
openai
python-dotenv
faiss-cpu
langchain
PyPDF2
python-docx
sentence-transformers
chromadb
